requirements.txt
numpy
scipy
matplotlib
scikit-image
Pillow
opencv-python
exifread
tqdm
gco-wrapper
pytest
